Change ilength in psf_bin_work to long

+++ MHD/C_src/KEMO_GL/set_map_isolines_to_buf.c
@@ -88,7 +88,7 @@ long set_map_PSF_isoline_to_buf(const long ist_patch,
                                 struct psf_data *psf_s, struct psf_menu_val *psf_m,
                                 struct gl_strided_buffer *psf_buf){
 	double dub_r;
-	int inum_patch = ist_patch;
+	long inum_patch = ist_patch;
 	if(psf_m->draw_psf_grid  != 0){
 		psf_m->ist_positive_line = find_start_positive_lines(psf_m->n_isoline, 
 								psf_m->cmap_psf_comp[psf_m->icomp_draw_psf]);
